fix: validate task timestamps strictly
