+++ apps/web/lib/utils.ts
@@ -1,6 +1,7 @@
 import { clsx, type ClassValue } from "clsx";
 import { twMerge } from "tailwind-merge";
 import { t } from "@/lib/i18n";
+import { parseStrictRfc3339Timestamp } from "@/lib/utils/strict-timestamp";
 
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
@@ -205,8 +206,8 @@ export const DEFAULT_LOCAL_EXECUTOR_TYPE = "worktree";
  * @returns Formatted relative time string
  */
 export function formatRelativeTime(dateString: string): string {
+  if (parseStrictRfc3339Timestamp(dateString) === null) return "";
   const date = new Date(dateString);
-  if (Number.isNaN(date.getTime())) return "";
   const now = new Date();
   const diffMs = now.getTime() - date.getTime();
   const diffSec = Math.floor(diffMs / 1000);
